Disable log to stdout from test runs #skip_bins
