+++ .gitignore
@@ -24,3 +24,4 @@ resources/models
 Hello-World_coding-example/build
 .pyenv
 .vscode
+test/OSC-ALKS-scenarios
